Add public identifiers to messages. Give "mine" a CSS class
Each message that is broadcast now gets an identifier for the user who
sent it. The identifier is a function of the message's sender id and
sender name, but does not leak the sender id (which might be a security
issue).

By comparing the current user's identifier with the one for an incoming
message, we can decide whether the message belongs to the current user
or not, and style "my" messages differently from "their" messages.