Fix two JS bugs
First, only bind event listener once on the close chat button, not every
time we start a session.

Second, don't add incoming messages to the DOM if the conversation has
already ended - eg, if the user closed it, they will still get a system
message saying chat ended, but should ignore it. If it's added to the
message list, it will just be hanging around there the next time they
open a conversation, and it will look like the new conversation has
ended.