Add concept of reps owning conversations
When the first rep joins a conversation, he/she will own it. We assign
a default id to reps with nil ids so that we can distinguish, by looking
at the owned_by_id column, from those that are not owned and those that
are owned by an anonymous rep.

Having ownership will let us provide better rep UI.